tour-operator-changes/toc-service/requirements.txt
flask
flask-login
flask-sqlalchemy
flask[async]
psycopg2-binary
